requirements.txt
Flask==3.0.0
gunicorn==21.2.0
tensorflow==2.15.0
tensorflow_macos==2.15.0
numpy==1.26.2
pandas==2.1.3
pillow==10.1.0